Rename endpoint_info to attrs.
