+++ iroh-dns/src/lib.rs
@@ -3,7 +3,7 @@
 //! This crate contains the core types for publishing and resolving iroh endpoint
 //! information via DNS, using the [pkarr](https://pkarr.org) signed packet format.
 
-pub mod endpoint_info;
+pub mod attrs;
 pub mod pkarr;
 
 use std::sync::LazyLock;
